Updated models.py for Resource / File association

+++ main/models.py
@@ -1023,16 +1023,20 @@ class Media(Model, ModelDiffMixin):
 class Resource(Model):
     path = CharField(db_index=True, max_length=256)
     media = ManyToManyField(Media, related_name='resource_media')
+    generic_files = ManyToManyField(File, related_name='resource_files')
     bucket = ForeignKey(Bucket, on_delete=PROTECT, null=True, blank=True)
 
     @transaction.atomic
-    def add_resource(path_or_link, media):
+    def add_resource(path_or_link, media, generic_file):
         if os.path.islink(path_or_link):
             path = os.readlink(path_or_link)
         else:
             path = path_or_link
-        if media is None:
+        if media is None and generic_file is None:
             obj, created = Resource.objects.get_or_create(path=path, bucket=None)
+        elif media is None:
+            obj, created = Resource.objects.get_or_create(path=path, bucket=generic_file.project.bucket)
+            obj.generic_files.add(generic_file)
         else:
             obj, created = Resource.objects.get_or_create(path=path, bucket=media.project.bucket)
             obj.media.add(media)
@@ -1045,7 +1049,7 @@ class Resource(Model):
                 path=os.readlink(path_or_link)
                 os.remove(path_or_link)
         obj = Resource.objects.get(path=path)
-        if obj.media.all().count() == 0:
+        if obj.media.all().count() == 0 and obj.generic_files.all().count() == 0:
             logger.info(f"Deleting object {path}")
             obj.delete()
             tator_store = get_tator_store(obj.bucket)
@@ -1113,6 +1117,19 @@ def safe_delete(path):
         Resource.delete_resource(path)
     except:
         logger.warning(f"Could not remove {path}")
+        logger.warning(f"{traceback.format_exc()}")
+
+def drop_file_from_resource(path, generic_file):
+    """ Drops the specified generic file from the resource. This should be called when
+        removing a resource from a File object but the File object is
+        not being deleted.
+    """
+    try:
+        logger.info(f"Dropping file {generic_file} from resource {path}")
+        obj = Resource.objects.get(path=path)
+        obj.generic_files.remove(generic_file)
+    except:
+        logger.warning(f"Could not remove {generic_file} from {path}")
         logger.warning(f"{traceback.format_exc()}")
 
 def drop_media_from_resource(path, media):
@@ -1510,10 +1527,18 @@ class File(Model, ModelDiffMixin):
 @receiver(post_save, sender=File)
 def file_save(sender, instance, created, **kwargs):
     TatorSearch().create_document(instance)
+    if instance.path and created:
+        Resource.add_resource(instance.path, None, instance)
 
 @receiver(pre_delete, sender=File)
 def file_delete(sender, instance, **kwargs):
     TatorSearch().delete_document(instance)
+
+@receiver(post_delete, sender=File)
+def file_post_delete(sender, instance, **kwargs):
+    # Delete the path reference
+    if not instance.path is None:
+        safe_delete(instance.path)
 
 def type_to_obj(typeObj):
     """Returns a data object for a given type object"""
